girafe final

+++ app/Http/Controllers/AdvertController.php
@@ -13,10 +13,12 @@ class AdvertController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
+
+
     public function index()
     {
-        $adverts = Advert::all()->paginate(15);
-        return view('adverts.index','adverts');
+        $adverts = Advert::with('user')->paginate(5);
+        return view('pages.advert.index', compact('adverts'));
     }
 
     /**
@@ -26,84 +28,117 @@ class AdvertController extends Controller
      */
     public function create()
     {
-        return view('advert.create');
+        if (Auth::check()) {
+
+            $advert = new Advert;
+            $method = 'post';
+            $value = "Create";
+            return view('pages.advert.edit', compact('method', 'value', 'advert'));
+        } else return redirect()->route('index');
     }
 
     /**
      * Store a newly created resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
+     * @param \Illuminate\Http\Request $request
      * @return \Illuminate\Http\Response
      */
     public function store(Request $request)
     {
+
+
         $data = $request->validate([
             'title' => ['required', 'min:10'],
-            'body' =>['required','min:25','max:255'],
+            'description' => ['required', 'min:25', 'max:2500',],
+            'image_patch' => ['required', 'max:2500', 'mimes:jpg,jpeg,png']
 
-                        ]);
+        ]);
+        $imageName = time() . '-' . $request->title . '.' . $request->image_patch->extension();
+        $request->image_patch->move(storage_path('app/images'), $imageName);
+
         $data['user_id'] = Auth::id();
-        Advert::create($data);
+        $data['image_patch'] = 'images/' . $imageName;
+        $advert = Advert::create($data);
 
-        return redirect()->route('advert.view',Auth::id())
+        return redirect()->route('show', $advert->id)
             ->with('success', 'Product created successfully.');
     }
 
     /**
      * Display the specified resource.
      *
-     * @param  \App\Models\Advert  $advert
+     * @param \App\Models\Advert $advert
      * @return \Illuminate\Http\Response
      */
-    public function show(Advert $advert)
+    public function show($id)
     {
-       $advert=Advert::find($advert);
-       return view('advert.show',compact('advert'));
+
+        $advert = Advert::find($id);
+        return view('pages.advert.showOne', compact('advert'));
+
     }
 
     /**
      * Show the form for editing the specified resource.
      *
-     * @param  \App\Models\Advert  $advert
+     * @param \App\Models\Advert $advert
      * @return \Illuminate\Http\Response
      */
     public function edit(Advert $advert)
     {
-        return view('advert.edit',compact('advert'));
+        if ($this->authorize('update', $advert)) {
 
+            $method = 'PUT';
+            $value = "Edit";
+            return view('.pages.advert.edit', compact('advert', 'method', 'value'));
+        }
     }
 
     /**
      * Update the specified resource in storage.
      *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  \App\Models\Advert  $advert
+     * @param \Illuminate\Http\Request $request
+     * @param \App\Models\Advert $advert
      * @return \Illuminate\Http\Response
      */
     public function update(Request $request, Advert $advert)
     {
-        $data = $request->validate([
-            'title' => ['required', 'min:10'],
-            'body' =>['required','min:25','max:255'],
+        if ($this->authorize('update', $advert)) {
+            $data = $request->validate([
+                'title' => ['required', 'min:10'],
+                'description' => ['required', 'min:25', 'max:2500'],
 
-        ]);
-        $data['user_id'] = Auth::id();
-        Advert::update($data);
+            ]);
 
-        return redirect()->route('advert.view',Auth::id())
-            ->with('success', 'Product updated successfully.');
+            $data['user_id'] = Auth::id();
+            $advert->update($data);
+
+            return redirect()->route('show', $advert->id)
+                ->with('success', 'Advert updated successfully.');
+        }
     }
 
     /**
      * Remove the specified resource from storage.
      *
-     * @param  \App\Models\Advert  $advert
+     * @param \App\Models\Advert $advert
      * @return \Illuminate\Http\Response
      */
+
+
     public function destroy(Advert $advert)
     {
-       $advert->delete();
-        return redirect()->route('advert.index',Auth::id())
-            ->with('success', 'Product deleted successfully.');
+        if ($this->authorize('delete', $advert)) {
+            $advert->delete();
+            return redirect()->route('index', Auth::id())
+                ->with('success', 'Product deleted successfully.');
+        }
+    }
+
+    public function searchByUser($id)
+    {
+        $adverts = Advert::with('user')->where('user_id', $id)->paginate(5);
+        return view('pages.advert.index', compact('adverts'));
+
     }
 }
